Adjust file structure (move shared parts to the global insight components directory)
