+++ client/web/src/enterprise/insights/components/creation-ui-kit/CodeInsightDashboardsVisibility.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Alert } from '@sourcegraph/wildcard'
+import { Alert } from '@sourcegraph/wildcard/out/src'
 
 export interface CodeInsightDashboardsVisibilityProps extends React.HTMLAttributes<HTMLDivElement> {
     dashboardCount: number
