+++ client/web/src/enterprise/insights/components/creation-ui-kit/index.ts
@@ -3,5 +3,5 @@ export { default as styles } from './CreationUiKit.module.scss'
 export * from './live-preview'
 export { getSanitizedRepositories } from './sanitizers/repositories'
 
-export { CodeInsightDashboardsVisibility } from './code-insight-dashboards-visibility/CodeInsightDashboardsVisibility'
+export { CodeInsightDashboardsVisibility } from './CodeInsightDashboardsVisibility'
 export { CodeInsightTimeStepPicker } from './code-insight-time-step-picker/CodeInsightTimeStepPicker'
