+++ client/web/src/enterprise/insights/components/index.ts
@@ -1,1 +1,5 @@
 export { CodeInsightsIcon } from '../../../insights/Icons'
+
+export * from './insights-view-grid'
+export * from './views'
+export * from './trancated-text/TrancatedText'
