+++ client/web/src/enterprise/insights/components/insights-view-grid/components/BuiltInInsight.tsx
@@ -12,7 +12,6 @@ import { InsightContentType } from '../../../core/types/insight/common'
 import { useDeleteInsight } from '../../../hooks/use-delete-insight'
 import { LazyQueryStatus } from '../../../hooks/use-parallel-requests/use-parallel-request'
 import { useRemoveInsightFromDashboard } from '../../../hooks/use-remove-insight'
-import { DashboardInsightsContext } from '../../../pages/dashboards/dashboard-page/components/dashboards-content/components/dashboard-inisghts/DashboardInsightsContext'
 import { getTrackingTypeByInsightType, useCodeInsightViewPings } from '../../../pings'
 import {
     CategoricalBasedChartTypes,
@@ -28,6 +27,7 @@ import {
 import { useInsightData } from '../hooks/use-insight-data'
 
 import { InsightContextMenu } from './insight-context-menu/InsightContextMenu'
+import { InsightContext } from './InsightContext'
 
 interface BuiltInInsightProps extends TelemetryProps, React.HTMLAttributes<HTMLElement> {
     insight: SearchRuntimeBasedInsight | LangStatsInsight
@@ -46,7 +46,7 @@ interface BuiltInInsightProps extends TelemetryProps, React.HTMLAttributes<HTMLE
 export function BuiltInInsight(props: BuiltInInsightProps): React.ReactElement {
     const { insight, resizing, telemetryService, innerRef, ...otherProps } = props
     const { getBuiltInInsightData } = useContext(CodeInsightsBackendContext)
-    const { dashboard } = useContext(DashboardInsightsContext)
+    const { dashboard } = useContext(InsightContext)
 
     const insightCardReference = useRef<HTMLDivElement>(null)
     const mergedInsightCardReference = useMergeRefs([insightCardReference, innerRef])
