+++ client/web/src/enterprise/insights/components/insights-view-grid/components/InsightContext.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react'
+
+import { InsightDashboard } from '../../../core'
+
+export interface DashboardInsightsContextData {
+    dashboard: InsightDashboard | null
+}
+
+export const InsightContext = createContext<DashboardInsightsContextData>({
+    dashboard: null,
+})
