+++ client/web/src/enterprise/insights/components/insights-view-grid/index.ts
@@ -1,0 +1,2 @@
+export { SmartInsightsViewGrid } from './SmartInsightsViewGrid'
+export { InsightContext } from './components/InsightContext'
