+++ client/web/src/enterprise/insights/components/trancated-text/TrancatedText.tsx
