+++ client/web/src/enterprise/insights/components/trancated-text/TruncatedText.module.scss
