+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/dashboard-select/components/index.ts
@@ -1,4 +1,4 @@
 export { InsightsBadge } from './insights-badge/InsightsBadge'
 export { MenuButton } from './menu-button/MenuButton'
 export { SelectOption, SelectDashboardOption } from './select-option/SelectOption'
-export { TruncatedText } from './trancated-text/TrancatedText'
+export { TruncatedText } from '../../../../../../components/trancated-text/TrancatedText'
