+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/dashboards-content/components/dashboard-inisghts/DashboardInsights.tsx
@@ -3,11 +3,9 @@ import React, { useContext, useMemo } from 'react'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { LoadingSpinner, useObservable } from '@sourcegraph/wildcard'
 
-import { SmartInsightsViewGrid } from '../../../../../../../components/insights-view-grid/SmartInsightsViewGrid'
+import { SmartInsightsViewGrid, InsightContext } from '../../../../../../../components'
 import { CodeInsightsBackendContext, InsightDashboard } from '../../../../../../../core'
 import { EmptyInsightDashboard } from '../empty-insight-dashboard/EmptyInsightDashboard'
-
-import { DashboardInsightsContext } from './DashboardInsightsContext'
 
 interface DashboardInsightsProps extends TelemetryProps {
     dashboard: InsightDashboard
@@ -29,12 +27,12 @@ export const DashboardInsights: React.FunctionComponent<DashboardInsightsProps> 
     }
 
     return (
-        <DashboardInsightsContext.Provider value={{ dashboard }}>
+        <InsightContext.Provider value={{ dashboard }}>
             {insights.length > 0 ? (
                 <SmartInsightsViewGrid insights={insights} telemetryService={telemetryService} className={className} />
             ) : (
                 <EmptyInsightDashboard dashboard={dashboard} onAddInsight={onAddInsightRequest} />
             )}
-        </DashboardInsightsContext.Provider>
+        </InsightContext.Provider>
     )
 }
