--- client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/dashboards-content/components/dashboard-inisghts/DashboardInsightsContext.ts
@@ -1,13 +1,0 @@
-import { createContext } from 'react'
-
-import { InsightDashboard } from '../../../../../../../core/types'
-
-export interface DashboardInsightsContextData {
-    dashboard: InsightDashboard | null
-}
-
-export const DEFAULT_DASHBOARD_CONTEXT: DashboardInsightsContextData = {
-    dashboard: null,
-}
-
-export const DashboardInsightsContext = createContext<DashboardInsightsContextData>(DEFAULT_DASHBOARD_CONTEXT)
