+++ client/web/src/enterprise/insights/pages/insights/creation/lang-stats/components/live-preview-chart/LangStatsInsightLivePreview.tsx
@@ -3,6 +3,7 @@ import React, { useContext, useMemo } from 'react'
 import { ErrorAlert } from '@sourcegraph/branded/src/components/alerts'
 import { useDeepMemo } from '@sourcegraph/wildcard'
 
+import { CategoricalBasedChartTypes, CategoricalChart } from '../../../../../../components'
 import {
     LivePreviewBanner,
     LivePreviewBlurBackdrop,
@@ -13,7 +14,6 @@ import {
     useLivePreview,
     StateStatus,
 } from '../../../../../../components/creation-ui-kit'
-import { CategoricalBasedChartTypes, CategoricalChart } from '../../../../../../components/views'
 import { CodeInsightsBackendContext, CategoricalChartContent } from '../../../../../../core'
 
 import { DEFAULT_PREVIEW_MOCK } from './constants'
